feat: 下单前信息获取和下单功能

+++ src/main/java/com/cxsj1/homework/w5/model/Form.java
@@ -53,6 +53,6 @@ public class Form {
 
     public final static class UpdateOrderForm {
         public String serial;
-        public Short status;
+        public Integer status;
     }
 }
